// src/components/ItemDetail.d.ts
import React from 'react';
import { TriageItem } from '../types';
interface ItemDetailProps {
    item: TriageItem;
    onUpdateItem: (item: TriageItem) => void;
    debugMode: boolean;
}
declare const ItemDetail: React.FC<ItemDetailProps>;
export default ItemDetail;
//# sourceMappingURL=ItemDetail.d.ts.map